refactor(shared): move body fat formulas into @workspace/shared
The web Daily Check-In and the mobile measurements screen both need to
estimate body fat, and a second copy of the formulas would drift from the
first. Move them to the shared workspace package so there is one
implementation to keep correct.

The web service module stays as the existing import path, re-exporting
from shared, so `BodyFatAlgorithm` and every call site are unchanged.
