+++ shared/src/index.ts
@@ -119,6 +119,7 @@ export * from "./nutrients/drinkPresetCatalog.ts";
 export * from "./nutrients/macroPickerFields.ts";
 export * from "./nutrients/alcoholUnits.ts";
 export * from "./nutrients/caffeineKinetics.ts";
+export * from "./utils/bodyComposition.ts";
 export * from "./utils/timezone.ts";
 export * from "./utils/calendar.ts";
 export * from "./utils/entryTime.ts";
